@@ -23,9 +23,9 @@ from __future__ import annotations
 
 import itertools
 import pickle
+from collections.abc import Callable
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Callable
 
 import pyphen
 
